removed not needed abstract class on binary protocol
